KPoPGacha/requirements.txt
python-telegram-bot==20.7
requests==2.31.0
dotenv==1.0.1 